665 moved filtered_qa after apportionment
