+++ src/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.6"
+__version__ = "0.6.7"
