+++ src/developer_config.yaml
@@ -27,6 +27,7 @@ global:
   output_auto_outliers: False
   output_outlier_qa : False
   output_estimation_qa: False
+  output_status_filtered: False
   output_apportionment_qa: False
   output_long_form: False
   output_short_form: False
@@ -37,7 +38,6 @@ global:
   output_intram_by_itl1: False
   output_intram_by_civil_defence: False
   output_intram_by_sic: False
-  output_status_filtered: False
   output_fte_total_qa: False
 years:
   current_year: 2022 # TODO: put this in the userconfig
